server/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pytube==15.0.0
deepgram-sdk==3.1.0
pydantic==2.6.3
sse-starlette==2.0.0
aiofiles==23.2.1
python-multipart==0.0.9 